+++ phpBB2/album_showpage.php
@@ -480,18 +480,7 @@ if( !isset($_POST['comment']) && !isset($_POST['rate']) )
 			$poster_avatar = '';
 			if ( $commentrow[$i]['user_avatar_type'] && $commentrow[$i]['user_id'] != ANONYMOUS && $commentrow[$i]['user_allowavatar'] )
 			{
-				switch( $commentrow[$i]['user_avatar_type'] )
-				{
-					case USER_AVATAR_UPLOAD:
-						$poster_avatar = ( $board_config['allow_avatar_upload'] ) ? '<img src="' . $board_config['avatar_path'] . '/' . $commentrow[$i]['user_avatar'] . '" alt="" border="0" />' : '';
-						break;
-					case USER_AVATAR_REMOTE:
-						$poster_avatar = ( $board_config['allow_avatar_remote'] ) ? '<img src="' . htmlspecialchars($commentrow[$i]['user_avatar'], ENT_QUOTES, 'UTF-8') . '" alt="" border="0" />' : '';
-						break;
-					case USER_AVATAR_GALLERY:
-						$poster_avatar = ( $board_config['allow_avatar_local'] ) ? '<img src="' . $board_config['avatar_gallery_path'] . '/' . $commentrow[$i]['user_avatar'] . '" alt="" border="0" />' : '';
-						break;
-				}
+				$poster_avatar = phpbb_avatar_image($commentrow[$i]['user_avatar'], $commentrow[$i]['user_avatar_type']);
 			}
 
 			//rank & rank image