+++ phpBB2/groupcp.php
@@ -43,18 +43,7 @@ function generate_user_info(&$row, $date_format, $group_mod, &$from, &$posts, &$
 	$poster_avatar = '';
 	if ( !empty($row['user_avatar_type']) && $row['user_id'] != ANONYMOUS && $row['user_allowavatar'] )
 	{
-		switch( $row['user_avatar_type'] )
-		{
-			case USER_AVATAR_UPLOAD:
-				$poster_avatar = ( $board_config['allow_avatar_upload'] ) ? '<img src="' . $board_config['avatar_path'] . '/' . $row['user_avatar'] . '" alt="" border="0" />' : '';
-				break;
-			case USER_AVATAR_REMOTE:
-				$poster_avatar = ( $board_config['allow_avatar_remote'] ) ? '<img src="' . htmlspecialchars($row['user_avatar'], ENT_QUOTES, 'UTF-8') . '" alt="" border="0" />' : '';
-				break;
-			case USER_AVATAR_GALLERY:
-				$poster_avatar = ( $board_config['allow_avatar_local'] ) ? '<img src="' . $board_config['avatar_gallery_path'] . '/' . $row['user_avatar'] . '" alt="" border="0" />' : '';
-				break;
-		}
+		$poster_avatar = phpbb_avatar_image($row['user_avatar'], $row['user_avatar_type']);
 	}
 
 	if ( !empty($row['user_viewemail']) || $group_mod )